fix(portal): don't use Web functions from Domain
