+++ elixir/apps/domain/lib/domain/mailer.ex
@@ -93,4 +93,20 @@ defmodule Domain.Mailer do
     Email.new()
     |> Email.from({"Firezone Notifications", from_email})
   end
+
+  def url(path, params \\ %{}) do
+    Domain.Config.fetch_env!(:domain, :web_external_url)
+    |> URI.parse()
+    |> URI.append_path(path)
+    |> maybe_append_query(params)
+    |> URI.to_string()
+  end
+
+  defp maybe_append_query(uri, params) do
+    if Enum.empty?(params) do
+      uri
+    else
+      URI.append_query(uri, URI.encode_query(params))
+    end
+  end
 end
